Refactor Subscription classes
- Introduce BaseSubscription and inherit from it to create EventStoreSubscription and BrokerSubscription classes
